pyproject.toml
[tool.poetry]
name = "vlg-bot"
version = "0.1.0"
description = "Персональный бот для сервера VLG"
authors = ["you@example.com"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
py-cord = "^2.4.1"
aiohttp = "^3.8.6"
discord-py = "^2.5.2"
markdown = "^3.8.2"
weasyprint = "^65.1"
asyncpg = "^0.30.0"
flask = "^3.1.1"
pynacl = "^1.5.0"
python-levenshtein = "^0.27.1"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"







